Time: 4 ms (86.66%), Space: 18.2 MB (67.78%) - LeetHub

+++ 0662-maximum-width-of-binary-tree/0662-maximum-width-of-binary-tree.cpp
@@ -18,7 +18,7 @@ public:
         
         while(!q.empty()) {
             int count = q.size();
-            long long minval = -1;
+            int minval = -1;
             int maxval = -1;
             for(int i = 0; i < count; i++) {
                 TreeNode* node = q.front().first;
